requirements.txt
# Core dependencies for multimodal RAG system
numpy>=1.21.0
scipy>=1.7.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
Pillow>=9.0.0

# OCR and image processing
pytesseract>=0.3.10
opencv-python>=4.5.0

# Audio processing
openai-whisper>=20231117
librosa>=0.9.0
soundfile>=0.12.0

# Embeddings and ML
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
sentence-transformers>=2.2.0
transformers>=4.21.0
clip-by-openai>=1.0

# Vector indexing
faiss-cpu>=1.7.0

# LLM integration
llama-cpp-python>=0.2.0

# API framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Database and storage
# sqlite3 is built-in with Python
sqlalchemy>=2.0.0

# Utilities
tqdm>=4.64.0
click>=8.0.0
pyyaml>=6.0
python-multipart>=0.0.6

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=5.0.0
mypy>=1.0.0

# Optional GPU support (uncomment if using CUDA)
# faiss-gpu>=1.7.0
# torch>=1.12.0+cu116